refactor: order processors in refactoring
